Transform project to DDD+hexagonal design
Transformed project to DDD+hexagonal design. Now we will have
a folder app with 4 main folders:
- controller: contains the controllers, which are the entry points
of the application. In this case, since it will only be accessible
through a web app, we only have http controller.
- application: contains the commands/use cases that are available
for execution. Gets called by the controllers and executes the
commands on the repository. Also interacts with the external
ports/interfaces.
- domain: currently only contains the model of the application. In
this case, it is totally integrated with mongoose.
- infrastructure: implementation of the interfaces to access the
models and external libraries. In this case we have the methods
to save/update/remove data from Mongo database, and interaction
with email

Extras:
- Implemented hashed tokens so even if an attacker gains access
to the Database, the password change will be secured

TODO:
- Implement the sending of emails as a set of commands according to
what kind of email they send (sendVerifyEmail, sendVerifiedEmail,...).
This commands will be subscribed to their corresponding events and
wil call the Emailer application to send them. So the methods will
be removed from Emailer, which should only be an interface.
- Named routes, so we don't need to put the full path every time,
similar to what Django has if possible, that can be used also
inside the templates
- Create a base template for all the sites