stripe: stripe integration for payments
Integrate payments with Stripe. In this version, when a user
validates his email, an Stripe customer will be created with his
name and email.

Then the user will go through a trial period and it will be offered
to buy a subscription using Stripe checkout portal. The user can
manage his subscription through Stripe customer portal.

The application listens for changes in Stripe using a webhook
and will synchronize with it whenever a user subscribes/unsubscribes.

We will need to define the following environment variables:
STRIPE_SECRET_KEY, STRIPE_WEBHOOK_SECRET, PRODUCT_BASIC and
TRIAL_DAYS

Also, for the webhook to work we will need to call in another terminal:
stripe listen --forward-to localhost:5656/subscription/stripe/webhook

TODO: When the app is started, we should query Stripe to get the events
that we missed